etl/requirements.txt
annotated-types==0.6.0
async-timeout==4.0.3
backoff==2.2.1
certifi==2024.2.2
elastic-transport==8.12.0
elasticsearch==8.12.1
psycopg2-binary==2.9.9
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
python-dotenv==1.0.1
redis==5.0.2
typing_extensions==4.10.0
urllib3==2.2.1
